requirements.txt
fastmcp==3.2.3
pydantic>=2.7.0
uvicorn[standard]>=0.30.0
supabase>=2.4.0
httpx>=0.27.0
anthropic>=0.25.0

# Ingestion layer — T1 pipelines
sqlalchemy>=2.0.0
celery[redis]>=5.3.0
redis>=5.0.0
tenacity>=8.2.0
pyyaml>=6.0.1
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
google-generativeai>=0.7.0

# Optional — install separately when needed:
# playwright>=1.40.0     (Skatteverket, Kronofogden, Boverket, Diarium scrapers)
# playwright install chromium
